% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{append_query}
\alias{append_query}
\title{Append the query elements}
\usage{
append_query(x, elems)
}
\arguments{
\item{x}{character, the query string}

\item{elems}{character, the elements as a named list}
}
\value{
input x with token appended
}
\description{
Append the query elements
}
