% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{append_token}
\alias{append_token}
\title{Append the token to a query}
\usage{
append_token(x, token = pp_token())
}
\arguments{
\item{x}{character, the query string}

\item{token}{character, the token}
}
\value{
input x with token appended
}
\description{
Append the token to a query
}
