% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa.R
\name{extract_geometry}
\alias{extract_geometry}
\title{Extract sf geometry}
\usage{
extract_geometry(x)
}
\arguments{
\item{x}{list of the geojson section of a response}
}
\value{
NULL or geometry
}
\description{
Extract sf geometry
}
