% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pp_browse}
\alias{pp_browse}
\title{Browse an single WPDA}
\usage{
pp_browse(x = 555703527, base_url = "https://www.protectedplanet.net")
}
\arguments{
\item{x}{character, one wdpaid (numeric or character code)}

\item{base_url}{the base url for Prtected Planet}
}
\description{
Browse an single WPDA
}
