% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pp_get_wdpa}
\alias{pp_get_wdpa}
\title{GET WPDA identified by ID}
\usage{
pp_get_wdpa(x = "40366", form = c("list", "tibble", "sf")[3])
}
\arguments{
\item{x}{character, one or more wdpaid (numeric or character code)}

\item{form}{character, describes output format. One of 'list', 'tibble' or 'sf'}
}
\value{
list, tibble, sf  I \code{length(x) > 1} then a tibble (or sf) is returned
}
\description{
GET WPDA identified by ID
}
