% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pp_get_wdpaid}
\alias{pp_get_wdpaid}
\title{GET WPDA identified by ID}
\usage{
pp_get_wdpaid(x = "40366", form = c("list", "tibble", "sf")[2])
}
\arguments{
\item{x}{character, a wdpaid (numeric or charcater code)}

\item{form}{character, describes output format. One of 'list', 'tibble' or 'sf'}
}
\value{
list, tibble, sf or NULL
}
\description{
GET WPDA identified by ID
}
