% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pp_path}
\alias{pp_path}
\title{Retrieve the local path to the Protected Planet resources}
\usage{
pp_path(..., root = rappdirs::user_data_dir("ppapi"))
}
\arguments{
\item{...}{character, file/directory segments}

\item{root}{character, the root directory}
}
\description{
Retrieve the local path to the Protected Planet resources
}
