% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pp_token}
\alias{pp_token}
\title{Retrieve the Protected Planet token}
\usage{
pp_token(path = rappdirs::user_config_dir("ppapi"))
}
\arguments{
\item{path}{character, the user configuration directory}
}
\value{
the token as character
}
\description{
Retrieve the Protected Planet token
}
