% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pp_url}
\alias{pp_url}
\title{Build a query}
\usage{
pp_url(
  x = "40366",
  with_geometry = TRUE,
  base_url = "http://api.protectedplanet.net",
  version = "v3",
  space = c("countries", "protected_areas")[2],
  token = pp_token()
)
}
\arguments{
\item{x}{character identifier (either iso3 for a country or wdpa_id for a protected area)}

\item{base_url}{character, the base url}

\item{version}{character, the API version to query}

\item{with_gemoetry}{logical to include boundary coordinates}
}
\value{
character url query
}
\description{
An example for a single protected area might look like 
\code{http://api.protectedplanet.net/v3/protected_areas/40366?with_geometry=true&token=0123456789}
where '0123456789' is replaced with a valid token.
}
