% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa.R
\name{wdpa_as_tibble}
\alias{wdpa_as_tibble}
\title{Convert a list of wdpa response items to tibble}
\usage{
wdpa_as_tibble(
  x,
  keep = c("id", "name", "original_name", "wdpa_id", "marine", "reported_marine_area",
    "reported_area", "management_plan", "owner_type", "countries", "iucn_category",
    "designation", "no_take_status", "legal_status", "management_authority",
    "governance", "legal_status_updated_at", "geometry")
)
}
\arguments{
\item{x}{list of wdpa items for one wdpa record}

\item{keep}{character, vector of columns to keep}
}
\value{
tibble (one row)
}
\description{
Convert a list of wdpa response items to tibble
}
